Alhamdulillah! bookings all routes, controllers, servicess added. just some polish need

+++ src/config/db.ts
@@ -40,7 +40,7 @@ const initDB = async () => {
         vehicle_id INT REFERENCES vehicles(id) ON DELETE CASCADE,
         rent_start_date DATE NOT NULL,
         rent_end_date DATE NOT NULL CHECK (rent_end_date > rent_start_date),
-        total_price NUMERIC(10, 2) NOT NULL CHECK (total_price > 0),
+        total_price INT NOT NULL CHECK (total_price > 0),
         status VARCHAR(20) NOT NULL DEFAULT 'active',
         CONSTRAINT bookking_status CHECK (status IN ('active','cancelled','returned'))
         )
